+++ src/org/azentreprise/arionide/ui/core/opengl/Geometry.java
@@ -22,10 +22,10 @@ package org.azentreprise.arionide.ui.core.opengl;
 
 import java.util.List;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 public interface Geometry {
 	public void setGenerationSeed(long seed);
-	public List<WorldElement> getCollisions(Vector3f player);
+	public List<WorldElement> getCollisions(Vector3d player);
 	public List<WorldElement> getElements();
 }
