+++ src/org/azentreprise/arionide/ui/core/opengl/WorldElement.java
@@ -24,6 +24,7 @@ import java.util.Random;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 import org.joml.Vector4f;
 
@@ -31,16 +32,16 @@ public class WorldElement {
 	private static final Random rand = new Random();
 	
 	// "Random": this one is not uniform (more likely to become vertical)
-	public static final Supplier<Vector3f> RANDOM_GENERATOR = () -> new Vector3f(rand.nextFloat() - 0.5f, rand.nextFloat() * 2.0f, rand.nextFloat() - 0.5f);
+	public static final Supplier<Vector3d> RANDOM_GENERATOR = () -> new Vector3d(rand.nextDouble() - 0.5d, rand.nextDouble() * 2.0d, rand.nextDouble() - 0.5d);
 
-	private static Supplier<Vector3f> axisGenerator = RANDOM_GENERATOR;
-	private static Function<Vector3f, Vector3f> baseGenerator = RANDOM_GENERATOR.get()::cross;
+	private static Supplier<Vector3d> axisGenerator = RANDOM_GENERATOR;
+	private static Function<Vector3d, Vector3d> baseGenerator = RANDOM_GENERATOR.get()::cross;
 
-	protected static void setAxisGenerator(Supplier<Vector3f> generator) {
+	protected static void setAxisGenerator(Supplier<Vector3d> generator) {
 		axisGenerator = generator;
 	}
 	
-	protected static void setBaseGenerator(Function<Vector3f, Vector3f> generator) {
+	protected static void setBaseGenerator(Function<Vector3d, Vector3d> generator) {
 		baseGenerator = generator;
 	}
 	
@@ -51,17 +52,17 @@ public class WorldElement {
 	/* Data needed for tesselation, rendering and collision detection. */
 	
 	private final int id;
-	private final Vector3f center;
-	private final Vector3f axis;
-	private final Vector3f base;
-	private final float size;
+	private final Vector3d center;
+	private final Vector3d axis;
+	private final Vector3d base;
+	private final double size;
 	private final boolean accessAllowed;
 	
 	private String name;
 	private Vector4f color;
 	private Vector3f spotColor;
 	
-	protected WorldElement(int id, String name, Vector3f center, Vector4f color, Vector3f spotColor, float size, boolean accessAllowed) {
+	protected WorldElement(int id, String name, Vector3d center, Vector4f color, Vector3f spotColor, double size, boolean accessAllowed) {
 		this.id = id;
 		this.name = name;
 		this.center = center;
@@ -73,7 +74,7 @@ public class WorldElement {
 		this.accessAllowed = accessAllowed;
 	}
 	
-	protected boolean collidesWith(Vector3f object) {
+	protected boolean collidesWith(Vector3d object) {
 		return object.distance(this.center) <= this.size;
 	}
 	
@@ -81,23 +82,23 @@ public class WorldElement {
 		return this.id;
 	}
 	
-	public Vector3f getCenter() {
-		return new Vector3f(this.center);
+	public Vector3d getCenter() {
+		return new Vector3d(this.center);
 	}
 	
-	public Vector3f getBaseVector() {
-		return new Vector3f(this.base);
+	public Vector3d getBaseVector() {
+		return new Vector3d(this.base);
 	}
 	
-	public Vector3f getAxis() {
-		return new Vector3f(this.axis);
+	public Vector3d getAxis() {
+		return new Vector3d(this.axis);
 	}
 	
 	public Vector3f getSpotColor() {
 		return this.spotColor;
 	}
 	
-	public float getSize() {
+	public double getSize() {
 		return this.size;
 	}
